[FLINK-21475][decl-scheduler] Support StateWithExecutionGraph.suspend when EG has reached globally terminal state
Since we don't execute StateWithExecutionGraph.onGloballyTerminalState immediately if the EG reaches
a globally terminal state, the StateWithExecutionGraph needs to support #suspend calls when the EG is
in a globally terminal state. In order to do this, the commit changes the precondition in
StateWithExecutionGraph.suspend from EG.getState() == JobStatus.SUSPEND to EG.getState().isTerminalState()
after calling EG.suspend.

This closes #15059.