[FLINK-25431] Implement a file-based JobResultStore
This commit contains an implementation of the JobResultStore backed by a
file-system. By default, this filesystem is based on the settings for
high-availability.storageDir and the cluster-id, but can be overriden
with the job-result-store.storage-path setting.

Alongside this setting, it also introduces a new option
job-result-store.delete-on-commit, which determines whether the
underlying file for the job result entry is deleted when it is clean, or
whether it is renamed and persisted. By default, this option is true.